fix: improve display touch detection
This adds some improvements to the SeeedStudio driver header for
detecting display touch. We now don't forward positions outside of the
display's circle to LVGL, since the touch controller regularly returns
positions like (1, 0), which is not possible due to the display being
circular and this coordinate not actually bein reachable, or something
in the billions of pixels, which is obviously also not possible. By
filtering these positions, the behavior of LVGL when using the
touchscreen becomes much more fluid and less jittery.