fix: improve display accuracy
Touch accuracy seems to work significantly better when not debouncing
the touch release as aggresively as we've been and when the touch sensor
is the only one using the I2C bus. The latter doesn't seem to be easy to
solve without making a new PCB. With the former solved, we can also
remove some of the more advanced filtering introduced earlier.