feat: use SCD41's temperature measurement instead of the BME's
With the current PCB layout, the BME is placed directly next to the ESP,
which, due to the ESP getting warm when in use, skews the measurements.
The SCD41 is currently far enough away and seems to be somewhat
isolated, for it to produce much better values, even without any
software correction done.