fix(scd): don't call stop_periodic_measurement before starting a periodic measurement
This is certainly less safe, but each stop_periodic_measurement takes
500ms, which quickly adds up (we were calling it in the constructor and
before starting a measurement again).
I kinda think we might want to add a MOSFET or something to the PCB in
the future and just power-cycle the sensor, instead of waiting all that
time for it to reset itself.