feat(schematic): remove unnecessary BME688
The BME688 we previously had on the PCB has gotten kind of useless,
since the SCD41 already has a temperature and humidity sensor on board.
This means the BME could only give us pressure (which, as the user, you
can't really do much with in my opinion), or VOC gas measurements,
which, to really do anything with, we'd need to use that BSEC library,
which feels kind of sketchy with the way its licensed.