feat: power sensors down when entering deep sleep
This commit introduces a system for tasks to register themselves as
needing to do work in order to prepare for entering deep sleep. When
display inactivity is triggered, we then wait for those tasks to do
their work before entering deep sleep. This allows us to tell the
sensors to power down and thus draw less current while sleeping and
thereby not in use.