feat: compass
We now use the LIS2MDL magnetometer reading to display the current
compass heading of the device. This is currently very basic and relies
on the device being flat on the table, as there is no tilt compensation
in place. The device then calculates the compass heading and displays it
in a new page on the screen.

To improve the accuracy, a hard-iron offset has been
calculated by plotting the measurments from as many different angles as
possible and offsetting those points such that they form a sphere around
the origin.