feat(compass): show angular offset of device
The compass now has as small circle in the middle that moves depending
on the angle of the device. When the device is parallel to the ground,
the circle will be in the middle and will move outwards the more the
device is rotated. This currently does not show the actual angle.