feat: use NFC chip to make URL with current sensor values available
The Sensor Puck will now update the NFC chip every 30s with a URL that
contains the current sensor values. In the future, this will include
more data, such as a history of the values. This allows the user to see
the data on a larger screen (i.e. their smartphone), as well as
additional information (e.g. a history) that isn't really displayable on
the small screen of the puck.