fix: prevent race conditions when waking up due to the timer expiring
Previously, we would deliberately initialize the UI before checking
whether the timer expired to allow it to react and give visual feedback.
However, since the UI is initialized in the LVGL task, there was a
chance that we would dispatch the timer expired event before the UI was
ready to react to it. Now, the init_display() function, which starts the
LVGL task, waits for the task to signal readiness, before returning,
hopefully guaranteeing that the user gets visual feedback of the timer
expiration. Similarly, the buzzer task is now started before this call,
which should give it enough time to start as well and provide auditory
feedback.