fix: wait for deep sleep preparation before clearing the display
Previously, sometimes when waking the puck up, you would still see a
sliver of the previously displayed frame. My guess is that this is
because the LVGL task is pushing another frame to the display after we
have already cleared it in preparation for deep sleep, making it briefly
visible on wakeup. Since the LVGL task is integrated in the deep sleep
preparation mechanism, we just need to wait for that to finish, before
clearing the display to hopefully avoid this issue.