fix(ui): prevent UI from displaying incorrect values if none are available yet
If no measurement values are available yet, we would display the default
initialisation from Data, which is obviously not a correct value. Now,
we don't immediately update the UI, but instead use the event system to
react to new values being available and display placeholders that show
that no data is available yet while waiting.