fix(nfc): add history time offset to URL and minor fixes
This adds a time offset of the last history entry to the current time in
minutes to the NFC URL payload. This allows the UI to accurately display
history values in relation to the current time. After the last one, all
history entries have a constant time offset between each other.

NOTE: At the moment, the UI does not do anything special with the
history offset.