refactor: stop using ULP RISC-V and wake the main CPU for periodic checks
In the past, the experience with the ULP RISC-V coprocessor has not been
great. The build system feels sketchy (I cannot remove anything using
the volatile variables orelse the variable definitions are not available
in the main project anymore) and it just straight up stopped working a
bit ago when removing code that really shouldn't have anything to do
with it (as far as I can tell...).

Now, we wake the main CPU periodically to check the sensors, update the
history and NFC payload and go back to sleep, without turning on the
screen. If the user timer would expire before we wake up due to this
period, we wake up when the timer would expire instead. A flag in RTC
RAM is used to distinguish between what we need to do on boot.

This is certainly more energy intensive and probably decreases battery
life, but it seems to be way more reliable than the using the ULP
processor. I would prefer using it nonetheless, but I guess I don't
understand it well enough to make it work properly.