fix: wake up immediately when the user touches the screen, even while performing environment check
The periodic environment check we do to notify the user of bad CO2
levels can take up to 5-6 seconds every minute. This means there is a
chance that the user wants to wake the device up by touching the screen
while we are busy performing said check, which would cause confusion as
the device is not reacting. Now, the device checks whether the user has
touched the screen to be able to react, even while silently checking the
CO2 levels.

This works by registering an interrupt for the touch pin, which sends a
task notification. We then check whether this task notification is
present at crucial points and if it is, we abort the check and continue
booting fully.